+++ tests/__init__.py
@@ -1,1 +1,1 @@
-# Tests for recordsQL
+# Tests for recordsql